problem/requirements.txt
asgiref==3.2.10
Django==3.1
django-bootstrap-modal-forms==2.0.0
django-flatpickr==1.0.1
django-widget-tweaks==1.4.8
pytz==2020.1
sqlparse==0.3.1
